Demo-Python-Deployment-main/requirements_GPU.txt
--extra-index-url https://download.pytorch.org/whl/cu126
--extra-index-url https://www.paddlepaddle.org.cn/packages/stable/cu126/

faiss-cpu
fastapi
google-api-core
google-auth
google-cloud-vision
matplotlib
mediapipe==0.10.21
numpy==1.26.4
opencv-python==4.11.0.86
opencv-python-headless==4.11.0.86
# PaddlePaddle (platform-specific)
paddlepaddle-gpu @ https://paddle-whl.bj.bcebos.com/stable/cu126/paddlepaddle-gpu/paddlepaddle_gpu-3.0.0-cp312-cp312-manylinux1_x86_64.whl; sys_platform == "linux" and python_version == "3.12"
paddlepaddle-gpu @ https://paddle-whl.bj.bcebos.com/stable/cu126/paddlepaddle-gpu/paddlepaddle_gpu-3.0.0-cp312-cp312-win_amd64.whl ; sys_platform == "win32" and python_version == "3.12"

insightface==0.7.3
onnxruntime==1.22.0
paddleocr==2.10.0
pillow>=11.1.0
pydantic==2.10.6
qdrant-client
requests
scikit-image==0.25.2
scipy==1.15.2
simsimd
six
tiktoken
tokenizers
torch==2.6.0
torchvision==0.21.0
torchaudio==2.6.0
uvicorn
python-dotenv
jwt
